fix: change days number type
Refs: #1

+++ src/admin.rs
@@ -29,7 +29,7 @@ pub trait AdminModule: crate::config::ConfigModule + storage::StorageModule {
 
     // days and amount needed to be bounded
     #[endpoint(setLockPeriodsAndBonds)]
-    fn set_lock_periods_and_bonds(&self, args: MultiValueEncoded<MultiValue2<u8, BigUint>>) {
+    fn set_lock_periods_and_bonds(&self, args: MultiValueEncoded<MultiValue2<u16, BigUint>>) {
         only_privileged!(self, ERR_NOT_PRIVILEGED);
         for input in args.into_iter() {
             let (lock_period, bond) = input.into_tuple();
